+++ mbrc-core/Core/Commands/Requests/PlayerStateCommands/RequestAutoDj.cs
@@ -4,16 +4,17 @@ using MusicBeeRemote.Core.Events;
 using MusicBeeRemote.Core.Events.Status.Internal;
 using MusicBeeRemote.Core.Model.Entities;
 using MusicBeeRemote.Core.Network;
+using Newtonsoft.Json.Linq;
 using TinyMessenger;
 
-namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
+namespace MusicBeeRemote.Core.Commands.Requests.PlayerStateCommands
 {
-    public class RequestStop : LimitedCommand
+    public class RequestAutoDj : LimitedCommand
     {
         private readonly ITinyMessengerHub _hub;
         private readonly IPlayerApiAdapter _apiAdapter;
 
-        public RequestStop(ITinyMessengerHub hub, IPlayerApiAdapter apiAdapter)
+        public RequestAutoDj(ITinyMessengerHub hub, IPlayerApiAdapter apiAdapter)
         {
             _hub = hub;
             _apiAdapter = apiAdapter;
@@ -22,7 +23,7 @@ namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
         /// <inheritdoc />
         public override string Name()
         {
-            return "Player: Stop";
+            return "Player: AutoDJ";
         }
 
         public override void Execute(IEvent receivedEvent)
@@ -32,13 +33,21 @@ namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
                 throw new ArgumentNullException(nameof(receivedEvent));
             }
 
-            var message = new SocketMessage(Constants.PlayerStop, _apiAdapter.StopPlayback());
-            _hub.Publish(new PluginResponseAvailableEvent(message, receivedEvent.ConnectionId));
+            var isToggle = receivedEvent.Data is JToken token &&
+                           ((string)token).Equals("toggle", StringComparison.InvariantCultureIgnoreCase);
+
+            if (isToggle)
+            {
+                _apiAdapter.ToggleAutoDjLegacy();
+            }
+
+            var message = new SocketMessage(Constants.PlayerAutoDj, _apiAdapter.IsAutoDjEnabledLegacy());
+            _hub.Publish(new PluginResponseAvailableEvent(message));
         }
 
         protected override CommandPermissions GetPermissions()
         {
-            return CommandPermissions.StopPlayback;
+            return CommandPermissions.ChangeShuffle;
         }
     }
 }
