+++ mbrc-core/Core/Commands/Requests/PlayerStateCommands/RequestNextTrack.cs
@@ -6,22 +6,22 @@ using MusicBeeRemote.Core.Model.Entities;
 using MusicBeeRemote.Core.Network;
 using TinyMessenger;
 
-namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
+namespace MusicBeeRemote.Core.Commands.Requests.PlayerStateCommands
 {
-    public class RequestPreviousTrack : LimitedCommand
+    public class RequestNextTrack : LimitedCommand
     {
-        private readonly ITinyMessengerHub _hub;
         private readonly IPlayerApiAdapter _apiAdapter;
+        private readonly ITinyMessengerHub _hub;
 
-        public RequestPreviousTrack(ITinyMessengerHub hub, IPlayerApiAdapter apiAdapter)
+        public RequestNextTrack(IPlayerApiAdapter apiAdapter, ITinyMessengerHub hub)
         {
-            _hub = hub;
             _apiAdapter = apiAdapter;
+            _hub = hub;
         }
 
         public override string Name()
         {
-            return "Player: Play previous";
+            return "Player: Play next";
         }
 
         public override void Execute(IEvent receivedEvent)
@@ -31,13 +31,13 @@ namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
                 throw new ArgumentNullException(nameof(receivedEvent));
             }
 
-            var message = new SocketMessage(Constants.PlayerPrevious, _apiAdapter.PlayPrevious());
+            var message = new SocketMessage(Constants.PlayerNext, _apiAdapter.PlayNext());
             _hub.Publish(new PluginResponseAvailableEvent(message, receivedEvent.ConnectionId));
         }
 
         protected override CommandPermissions GetPermissions()
         {
-            return CommandPermissions.PlayPrevious;
+            return CommandPermissions.PlayNext;
         }
     }
 }
