+++ mbrc-core/Core/Commands/Requests/PlayerStateCommands/RequestPlayPause.cs
@@ -6,22 +6,23 @@ using MusicBeeRemote.Core.Model.Entities;
 using MusicBeeRemote.Core.Network;
 using TinyMessenger;
 
-namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
+namespace MusicBeeRemote.Core.Commands.Requests.PlayerStateCommands
 {
-    public class RequestNextTrack : LimitedCommand
+    public class RequestPlayPause : LimitedCommand
     {
-        private readonly IPlayerApiAdapter _apiAdapter;
         private readonly ITinyMessengerHub _hub;
+        private readonly IPlayerApiAdapter _apiAdapter;
 
-        public RequestNextTrack(IPlayerApiAdapter apiAdapter, ITinyMessengerHub hub)
+        public RequestPlayPause(ITinyMessengerHub hub, IPlayerApiAdapter apiAdapter)
         {
-            _apiAdapter = apiAdapter;
             _hub = hub;
+            _apiAdapter = apiAdapter;
         }
 
+        /// <inheritdoc />
         public override string Name()
         {
-            return "Player: Play next";
+            return "Player: Play/Pause";
         }
 
         public override void Execute(IEvent receivedEvent)
@@ -31,13 +32,14 @@ namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
                 throw new ArgumentNullException(nameof(receivedEvent));
             }
 
-            var message = new SocketMessage(Constants.PlayerNext, _apiAdapter.PlayNext());
+            var message = new SocketMessage(Constants.PlayerPause, _apiAdapter.PlayPause());
             _hub.Publish(new PluginResponseAvailableEvent(message, receivedEvent.ConnectionId));
         }
 
         protected override CommandPermissions GetPermissions()
         {
-            return CommandPermissions.PlayNext;
+            return CommandPermissions.StartPlayback |
+                   CommandPermissions.StopPlayback;
         }
     }
 }
