+++ mbrc-core/Core/Commands/Requests/PlayerStateCommands/RequestPlayerStatus.cs
@@ -4,17 +4,16 @@ using MusicBeeRemote.Core.Events;
 using MusicBeeRemote.Core.Events.Status.Internal;
 using MusicBeeRemote.Core.Model.Entities;
 using MusicBeeRemote.Core.Network;
-using Newtonsoft.Json.Linq;
 using TinyMessenger;
 
-namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
+namespace MusicBeeRemote.Core.Commands.Requests.PlayerStateCommands
 {
-    public class RequestScrobble : ICommand
+    public class RequestPlayerStatus : ICommand
     {
         private readonly ITinyMessengerHub _hub;
         private readonly IPlayerApiAdapter _apiAdapter;
 
-        public RequestScrobble(ITinyMessengerHub hub, IPlayerApiAdapter apiAdapter)
+        public RequestPlayerStatus(ITinyMessengerHub hub, IPlayerApiAdapter apiAdapter)
         {
             _hub = hub;
             _apiAdapter = apiAdapter;
@@ -27,14 +26,8 @@ namespace MusicBeeRemote.Core.Commands.Requests.PlayerState
                 throw new ArgumentNullException(nameof(receivedEvent));
             }
 
-            if (receivedEvent.Data is JToken token &&
-                ((string)token).Equals("toggle", StringComparison.InvariantCultureIgnoreCase))
-            {
-                _apiAdapter.ToggleScrobbling();
-            }
-
-            var message = new SocketMessage(Constants.PlayerScrobble, _apiAdapter.ScrobblingEnabled());
-            _hub.Publish(new PluginResponseAvailableEvent(message));
+            var statusMessage = new SocketMessage(Constants.PlayerStatus, _apiAdapter.GetStatus());
+            _hub.Publish(new PluginResponseAvailableEvent(statusMessage, receivedEvent.ConnectionId));
         }
     }
 }
