+++ mbrc-core/mbrc-core.csproj
@@ -124,19 +124,19 @@
     <Compile Include="Core\Commands\Requests\Output\RequestPlayerOutputSwitch.cs" />
     <Compile Include="Core\Commands\Requests\PaginatedData.cs" />
     <Compile Include="Core\Commands\Requests\PaginatedDataCommand.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestAutoDj.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestMute.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestNextTrack.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestPause.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestPlay.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestPlayerStatus.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestPlayPause.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestPreviousTrack.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestRepeat.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestScrobble.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestShuffle.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestStop.cs" />
-    <Compile Include="Core\Commands\Requests\PlayerState\RequestVolume.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestAutoDj.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestMute.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestNextTrack.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestPause.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestPlay.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestPlayerStatus.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestPlayPause.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestPreviousTrack.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestRepeat.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestScrobble.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestShuffle.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestStop.cs" />
+    <Compile Include="Core\Commands\Requests\PlayerStateCommands\RequestVolume.cs" />
     <Compile Include="Core\Commands\Requests\PlayingTrack\RequestPlaybackPosition.cs" />
     <Compile Include="Core\Commands\Requests\PlayingTrack\RequestRating.cs" />
     <Compile Include="Core\Commands\Requests\PlayingTrack\RequestSongInfo.cs" />
